@@ -7,52 +7,56 @@
   "author": "Simon Kostin <s-kostin@t-online.de>",
   "license": "MIT",
   "dependencies": {
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.60",
-    "@types/react-custom-scrollbars": "^4.0.8",
-    "clsx": "^1.1.1",
-    "core-js": "^3.16.1",
-    "react": "^17.0.2",
-    "react-custom-scrollbars": "^4.2.1",
-    "react-dom": "^17.0.2",
-    "react-particles-js": "^3.5.3",
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
+    "@mui/icons-material": "^5.8.0",
+    "@mui/lab": "^5.0.0-alpha.82",
+    "@mui/material": "^5.8.0",
+    "@mui/styles": "^5.8.0",
+    "@types/react-custom-scrollbars": "^4.0.10",
+    "core-js": "^3.22.5",
+    "react": "^18.1.0",
+    "react-custom-scrollbars-2": "^4.4.0",
+    "react-dom": "^18.1.0",
+    "react-tsparticles": "^2.0.6",
     "regenerator-runtime": "^0.13.9"
   },
   "scripts": {
-    "serve": "webpack serve --mode development --color --hot --progress --config ./webpack.dev.js",
-    "serve:network": "webpack serve --mode development --color --hot --progress --config ./webpack.dev.js --host 0.0.0.0"
+    "serve": "webpack serve --mode development --color --hot --progress --config ./webpack/webpack.dev.js",
+    "serve:network": "webpack serve --mode development --color --hot --progress --config ./webpack/webpack.dev.js --host 0.0.0.0",
+    "build": "webpack --config ./webpack/webpack.prod.js"
   },
   "devDependencies": {
-    "@babel/core": "^7.15.0",
-    "@babel/preset-env": "^7.15.0",
-    "@babel/preset-react": "^7.14.5",
-    "@babel/preset-typescript": "^7.15.0",
-    "@types/react": "^17.0.16",
-    "@types/react-dom": "^17.0.9",
-    "@typescript-eslint/eslint-plugin": "^4.29.1",
-    "@typescript-eslint/parser": "^4.29.1",
-    "babel-loader": "^8.2.2",
-    "css-loader": "^6.2.0",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-react": "^7.24.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "@babel/core": "^7.18.0",
+    "@babel/preset-env": "^7.18.0",
+    "@babel/preset-react": "^7.17.12",
+    "@babel/preset-typescript": "^7.17.12",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.4",
+    "@typescript-eslint/eslint-plugin": "^5.25.0",
+    "@typescript-eslint/parser": "^5.25.0",
+    "babel-loader": "^8.2.5",
+    "css-loader": "^6.7.1",
+    "eslint": "^8.16.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.30.0",
+    "eslint-plugin-react-hooks": "^4.5.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "file-loader": "^6.2.0",
-    "html-loader": "^2.1.2",
-    "html-webpack-plugin": "^5.3.2",
-    "node-sass": "^6.0.1",
-    "prettier": "^2.3.2",
-    "sass-loader": "^12.1.0",
-    "source-map-loader": "^3.0.0",
-    "style-loader": "^3.2.1",
-    "ts-loader": "^9.2.5",
-    "tsparticles": "^1.33.3",
-    "typescript": "^4.3.5",
-    "webpack": "^5.50.0",
-    "webpack-cli": "^4.7.2",
-    "webpack-dev-server": "^3.11.2"
+    "html-loader": "^3.1.0",
+    "html-webpack-plugin": "^5.5.0",
+    "node-sass": "^7.0.1",
+    "prettier": "^2.6.2",
+    "sass-loader": "^13.0.0",
+    "source-map-loader": "^3.0.1",
+    "style-loader": "^3.3.1",
+    "ts-loader": "^9.3.0",
+    "tsparticles": "^2.0.6",
+    "typescript": "^4.6.4",
+    "webpack": "^5.72.1",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.9.0",
+    "webpack-merge": "^5.8.0"
   }
 }
