@@ -16,12 +16,12 @@ import {
   CustomTimelineOppositeContent,
   CustomTimelineSeparator,
 } from '@customizations/timeline.customization';
-import { makeStyles } from '@material-ui/core';
-import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
-import { TimelineItem } from '@material-ui/lab';
+import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
+import { TimelineItem } from '@mui/lab';
+import makeStyles from '@mui/styles/makeStyles';
 import clsx from 'clsx';
 import React, { ReactElement } from 'react';
-import { Scrollbars } from 'react-custom-scrollbars';
+import { Scrollbars } from 'react-custom-scrollbars-2';
 
 const useStyles = makeStyles(() => ({
   educationComponent: {
@@ -56,7 +56,7 @@ export const EducationComponent = (): ReactElement => {
   return (
     <Scrollbars hideTracksWhenNotNeeded={true}>
       <div className={classes.educationComponent}>
-        <CustomTimeline align="alternate">
+        <CustomTimeline position="alternate">
           <TimelineItem>
             <CustomTimelineOppositeContent>
               2019 - Today
