@@ -1,9 +1,9 @@
-import { IconButton, makeStyles, Theme } from '@material-ui/core';
-import FacebookIcon from '@material-ui/icons/Facebook';
-import GitHubIcon from '@material-ui/icons/GitHub';
-import LinkedInIcon from '@material-ui/icons/LinkedIn';
-import MailIcon from '@material-ui/icons/Mail';
-import SettingsIcon from '@material-ui/icons/Settings';
+import { IconButton, Theme } from '@mui/material';
+import makeStyles from '@mui/styles/makeStyles';
+import FacebookIcon from '@mui/icons-material/Facebook';
+import GitHubIcon from '@mui/icons-material/GitHub';
+import LinkedInIcon from '@mui/icons-material/LinkedIn';
+import MailIcon from '@mui/icons-material/Mail';
 import React, { ReactElement } from 'react';
 
 interface IFooterProps {}
@@ -27,16 +27,16 @@ export const Footer = (props: IFooterProps): ReactElement => {
   const classes = useStyles();
   return (
     <div className={classes.footer}>
-      <IconButton color="inherit">
+      <IconButton color="inherit" size="large">
         <GitHubIcon fontSize="inherit" />
       </IconButton>
-      <IconButton color="inherit">
+      <IconButton color="inherit" size="large">
         <FacebookIcon fontSize="inherit" />
       </IconButton>
-      <IconButton color="inherit">
+      <IconButton color="inherit" size="large">
         <LinkedInIcon fontSize="inherit" />
       </IconButton>
-      <IconButton color="inherit">
+      <IconButton color="inherit" size="large">
         <MailIcon fontSize="inherit" />
       </IconButton>
     </div>
