@@ -1,5 +1,5 @@
-import { LinearProgress, Theme } from '@material-ui/core';
-import { withStyles } from '@material-ui/styles';
+import { LinearProgress, Theme } from '@mui/material';
+import { withStyles } from '@mui/styles';
 
 export const CustomProgress = withStyles((theme: Theme) => ({
   root: {
@@ -9,7 +9,7 @@ export const CustomProgress = withStyles((theme: Theme) => ({
   },
   colorPrimary: {
     backgroundColor:
-      theme.palette.grey[theme.palette.type === 'light' ? 200 : 700],
+      theme.palette.grey[theme.palette.mode === 'light' ? 200 : 700],
   },
   bar: {
     borderRadius: 5,
