@@ -1,20 +1,21 @@
-import {Theme, ThemeOptions} from '@material-ui/core/styles/createTheme'
-declare module '@material-ui/core/styles/createTheme' {
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+import { Theme, ThemeOptions } from '@mui/material/styles';
+declare module '@mui/material/styles/createTheme' {
   interface Theme {
     custom: {
-      topHeight: number,
-      bottomHeight: number,
-      paperLight: string,
-      backgroundGradient: string,
-    }
+      topHeight: number;
+      bottomHeight: number;
+      paperLight: string;
+      backgroundGradient: string;
+    };
   }
   // allow configuration using `createMuiTheme`
   interface ThemeOptions {
     custom?: {
-      topHeight?: number,
-      bottomHeight?: number,
-      paperLight?: string,
-      backgroundGradient?: string,
-    }
+      topHeight?: number;
+      bottomHeight?: number;
+      paperLight?: string;
+      backgroundGradient?: string;
+    };
   }
 }
